requirements.txt
bcrypt==4.1.1
blinker==1.7.0
cassandra-driver==3.28.0
click==8.1.7
Flask==3.0.0
geomet==0.2.1.post1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
six==1.16.0
Werkzeug==3.0.1
